feat: Add tags.csv to export tarball
