+++ xmodule/modulestore/xml_exporter.py
@@ -12,6 +12,7 @@ import lxml.etree
 from fs.osfs import OSFS
 from opaque_keys.edx.locator import CourseLocator, LibraryLocator
 from xblock.fields import Reference, ReferenceList, ReferenceValueDict, Scope
+from openedx.core.djangoapps.content_tagging.api import export_tags_in_csv_file
 
 from xmodule.assetstore import AssetMetadata
 from xmodule.contentstore.content import StaticContent
@@ -281,6 +282,8 @@ class CourseExportManager(ExportManager):
             course_policy.write(dumps(policy, cls=EdxJSONEncoder, sort_keys=True, indent=4).encode('utf-8'))
 
         _export_drafts(self.modulestore, self.courselike_key, export_fs, xml_centric_courselike_key)
+
+        export_tags_in_csv_file(str(self.courselike_key), export_fs, 'tags.csv')
 
 
 class LibraryExportManager(ExportManager):
